# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Kosha REST serice for exposing many freshdesk features as REST APIs with better consistency, observability etc",
        "title": "Freshdesk Connector API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "eti@cisco.io"
        },
        "version": "2.0"
    },
    "host": "localhost:8015",
    "basePath": "/",
    "paths": {
...
        "/api/v1/tickets": {
            "get": {
                "description": "List all tickets logged in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Get all tickets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Ticket"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create single ticket in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Create new ticket",
                "parameters": [
                    {
                        "description": "Enter ticket properties",
                        "name": "text",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.Ticket"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/v1/tickets/{id}": {
            "get": {
                "description": "List single ticket based on the ticket id logged in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Get single ticket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enter ticket id",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Ticket"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "permission denied",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete single ticket based on the ticket id logged in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Delete single ticket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enter ticket id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Ticket": {
            "type": "object",
            "properties": {
                "associated_tickets_count": {},
                "association_type": {},
                "attachments": {
                    "type": "array",
                    "items": {}
                },
                "cc_emails": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "company": {
                    "type": "object"
                },
                "company_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "custom_fields": {
                    "type": "object"
                },
                "description": {
                    "type": "string"
                },
                "description_text": {
                    "type": "string"
                },
                "due_by": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_config_id": {
                    "type": "integer"
                },
                "fr_due_by": {
                    "type": "string"
                },
                "fr_escalated": {
                    "type": "boolean"
                },
                "fwd_emails": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "group_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_escalated": {
                    "type": "boolean"
                },
                "nr_due_by": {},
                "nr_escalated": {
                    "type": "boolean"
                },
                "priority": {},
                "product_id": {
                    "type": "integer"
                },
                "reply_cc_emails": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "requester": {
                    "type": "object",
                    "properties": {
                        "email": {
                            "type": "string"
                        },
                        "id": {
                            "type": "integer"
                        },
                        "mobile": {
                            "type": "integer"
                        },
                        "name": {
                            "type": "string"
                        },
                        "phone": {
                            "type": "string"
                        }
                    }
                },
                "requester_id": {
                    "type": "integer"
                },
                "responder_id": {
                    "type": "integer"
                },
                "source": {},
                "source_additional_info": {},
                "spam": {
                    "type": "boolean"
                },
                "stats": {
                    "type": "object",
                    "properties": {
                        "agent_responded_at": {},
                        "closed_at": {},
                        "first_responded_at": {},
                        "pending_since": {},
                        "reopened_at": {},
                        "requester_responded_at": {},
                        "resolved_at": {},
                        "status_updated_at": {
                            "type": "string"
                        }
                    }
                },
                "status": {},
                "subject": {
                    "type": "string"
                },
                "support_email": {},
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ticket_cc_emails": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "to_emails": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "twitter_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}